Define JAVA_TOOL_OPTIONS as a system-wide variable when required
As of JDK 8u331 / JDK 11.0.15+8 : The parsing of URLs in the LDAP, DNS, and RMI built-in JNDI providers has been made more strict, which effectively breaks nodetool in scylla-tools releases prior to when CASSANDRA-17581 was fixed. Users running an earlier release of Scylla will fail on nodetool with the following message:

```shell
nodetool: Failed to connect to '127.0.0.1:7199' - URISyntaxException: 'Malformed IPv6 address at index 7: rmi://[127.0.0.1]:7199'.
```

This affects all distributions using a recent OpenJDK release: Debian & Ubuntu (OpenJDK 11) and Centos 7 (OpenJDK 8) have been tested and are all affected.

The known workarounds to this are:
 - Append the `-Dcom.sun.jndi.rmiURLParsing="legacy"` option to nodetool command line
 - Call nodetool with `-h ::FFFF:127.0.0.1` on the command line
 - Revert to an older OpenJDK release known not to be affected by this change.

We decided to go with the former, as the 2nd option seemed a burden to require the user to type everytime, and the latter is not a good security practice in general.

This commit will therefore append JAVA_TOOL_OPTIONS="-Dcom.sun.jndi.rmiURLParsing=legacy" to the host OS /etc/environment file, which allows any user to run nodetool commands when a recent OpenJDK release is installed AND when running an older scylla-tools version.

As this corrects a "transitory" situation, it is placed as a separate commit so that it can easily be reverted later, should it be needed, when all supported Scylla releases are known to have the fix implemented.