Node Role - Proper Generate a CA certificate
Our previous implementation of ssl.yml incorrectly generated a self-signed certificate mistakenly named as a "CA".

As it turns out, a CA certificate needs to have some specific characteristics in order to be considered a proper Certificate Authority.
Due to that, the CA certificate couldn't be used by clients to properly authenticate against the nodes, in such a way that clients were
required to skip TLS verification in order to connect.

The way a CA certificate is now corrected and follows Ansible guidelines as listed in <https://docs.ansible.com/ansible/latest/collections/community/crypto/docsite/guide_ownca.html>