Node role - Correct cqlshrc when using TLS
The previous cqlshrc.j2 template didn't made too much sense. In particular, due to the fact that the CA was incorrectly generated.
We now properly specify the CA path and ensure we verify its validity.

Then we specify that we want TLSv1.2 protocol as Scylla correctly turns off TLS1.0 and only very few ciphers are enabled for TLS1.1.

Finally we remove the need to import the certificates from all nodes within cqlshrc - as now we have a trusted (and valid!) CA.

Notice that users connecting to Scylla via cqlsh are required to issue `cqlsh --ssl` in order to use SSL, otherwise cqlsh ignores the config file and tries to establish a plain-text connection which - of course - fails.