ansible-scylla-node: Verify if auth is enabled and adjust system_auth ks if necessary
By default, system_auth keyspace uses simple strategy and replication factor 1
and these settings are not recommended for production since it don't guarantee
that the information in this keyspace will be highly available for the cluster.
This patch gives the option to update the replication settings for the system_auth keyspace
automatically whenever authentication is enabled.

To make this possible, this patch adds the following variables:
adjust_system_auth_replication -> a flag indicating if the system_auth replication should be adjusted automatically
system_auth_rf -> The replication factor to be used
system_auth_replication_strategy -> The replication strategy to be used
cql_username -> A cql user with permissions to update replication settings for the system_auth ks
cql_password -> The password for the cql user provided
cleanup_timeout_seconds -> The maximum time the playbook should wait for the cleanup to finish