ansible-scylla-node: Add a repair task
This patch adds a repair task which allows you to repair a specific keyspace
by using Scylla rest API served from scylla_api_address to trigger a repair of each one
of the primary token ranges belonging to the current node.
It also adds a repair_timeout_seconds variable, in order to control the maximum time
the playbook should wait for the repair to finish.