rolling_restart.yml: wait for a CQL port to become operational before continuing
It's not enough to wait for a gossip state (UN in the 'nodetool status' output).
We need to wait for a scylla to become fully operational and this is
when both a gossip state is UN and scylla listens to all relevant ports.
And and CQL protocol is initialized the last (naturally).

So, let's wait for a CQL before checking the gossip state.

Fixes #26

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>