ansible-scylla-node: Allow user to provide their own certificates
Today, if someone wants to use their own node certificates or sign the node certificates
with their own CA this is not possible, even though it's described on the documentation
as possible. This PR adds this functionality following part of the instructions already
described in the documentation and also makes a few adjustments to the documentation.