ansible-scylla-node: Parametrize localhost certificate paths
Currently, in order to use your own certificates, they should always be placed in the
same directory in which you're running ansible.
Since this might not always be the desired behavior this patch adds a variable
scylla_ssl.localhost_cert_path which defaults to the inventory directory but allows you
to choose any other directory as well.