ansible-scylla-monitoring: Pass cql credentials to scylla-monitoring
This change is necessary to allow the usage of scylla cql plugin on grafana
when authentication is enabled on scylla.
This patch adds a variable scylla_monitoring_cql_default_user to the role and expects
that if auth is enabled on scylla, the username inside this variable will be defined
in the inventory with its respective password under the section [cql_credentials].
E.g (assuming scylla_monitoring_cql_default_user: 'scylla_cql_monitor'):
[cql_credentials]
scylla_cql_monitor=mysecretpassword

This patch set the env variables SCYLLA_USER and SCYLLA_PSSWD before running
start-all.sh, taking advantage of the fact that, starting from version 4.0.0,
scylla-monitoring uses these variables to configure scylla cql plugin.

This patch doesn't cover the non-docker Monitoring installation, which will be
fixed in a follow-up patch.