ansible-scylla-node: Verify if audit is enabled and adjust audit ks if necessary
By default, if audit is enabled, audit keyspace uses simple strategy and replication factor 1
and these settings are not recommended for production since it don't guarantee
that the information in this keyspace will be highly available for the cluster.
This patch gives the option to update the replication settings for the audit keyspace
automatically whenever audit is enabled.

To make this possible, this patch adds the following variables:
adjust_audit_replication -> a flag indicating if the audit replication should be adjusted automatically
audit_rf -> The replication factor to be used
audit_replication_strategy -> The replication strategy to be used