ansible-scylla-node: Replace cql_username/cql_password with scylla_admin_default_user
This patch adds a variable scylla_admin_default_user to the role and expects
that if adjust_replication flags are set to True, the username inside this variable will be defined
in the inventory with its respective password under the section [cql_credentials].
E.g (assuming scylla_admin_default_user: 'scylla_admin'):
[cql_credentials]
scylla_admin=mysecretpassword

If the scylla_admin_default_user is not defined, the role will try to
use 'cassandra' as username/password.
This patch also removes the variables cql_username/cql_password, since they are not
necessary anymore.