ansible-scylla-node: Convert io_properties variable to yaml
Currently, if we manually set io_properties, ansible will, by default, read it as json
instead of as yaml and as a result, io_properties.yaml will have a json array
as content. This patch fixes this by converting the content in io_properties var to
yaml before copying to the file.