manage_users.yml: Add playbook to create cql users
This patch adds a playbook which allows the creation of cql users and deletes the default cassandra user as recommended in scylla's
documentation -> {https://docs.scylladb.com/stable/operating-scylla/security/enable-authorization.html#enabling-authorization}.
By default this playbook will add the users scylla_admin (superuser), scylla_cql_monitor and customer_admin,
but you can change the list of users and permissions by overriding the variable users.
Every user defined in this playbook must be also defined in the inventory file under the section [cql_credentials] followed
by its password.
E.g.:
[cql_credentials]
scylla_admin=password1
scylla_cql_monitor=password2
customer_admin=password3

In order to connect via cqlsh and update the credentials, this playbook will try to get the rpc_address on scylla.yaml.
If this value is not set, then it'll be assumed that localhost is being used, since this is the default value on scylla and the
playbook will try to find out the IP address by getting the IPv4 for the default network interface.
If you want, you can override the variables scylla_nic and scylla_nic_ipv4_addr to choose the nic and ipv4 address which you'd like
the playbook to use in case rpc is not set.
If rpc_address is set to 0.0.0.0, this playbook will use broadcast_rpc_address.

Operations allowed:
  * Addition of users
  * Addition of permissions for any user (Removing permissions isn't supported)
  * Changing password for any user except scylla_admin

Usage:
  ansible-playbook -i inventory.ini manage_users.yaml [-e "@users.yml"]