ansible-scylla-monitoring: Move scylla_servers.yml generation to the monitoring role
Currently, scylla_servers generation is done in the node role and this creates
an unnecessary dependency from the monitoring role on the node role, since
this file is useful only for the former.
This patch fixes this by moving the files related to the generation of this file
to the monitoring role. It also renames monitoring_config.yml to generate_scylla_servers.yml
and adds the variables scylla_cluster_name, scylla_api_address and scylla_api_port to the monitoring role,
since these variables are dependencies of the moved task.