manage_user.yml: Use import_task instead of include_tasks for creating admin_user
Passing ignore errors to the task included works only with import_tasks.
This patch replaces include_tasks with import_tasks in order for the
ignore_errors to work. This ignore_errors is necessary because it's ok
for the task of creating scylla_admin to fail if scylla_admin already
exists.