manage_users.yml: Add ALTER permission on ALL ROLES to customer_admin
Currently, the customer_admin user is not able to change the password for other users.
This patch fixes this by giving ALTER permission on ALL roles to this user.
This is not ideal, since now the customer_admin user will be able to change passwords
even for superusers, but it's the only option we have until the following issues are fixed:

- https://github.com/scylladb/scylladb/issues/14277
- https://github.com/scylladb/scylladb/issues/14279