manage_users.yml: Divide the former variable users between internal_users and customer_users
The idea of this patch is that internal_users should be used for users for which
we want to control their state. An example would be a scylla_cql_monitor user. This
user is used only for monitoring and it makes sense to always have the same set of
permissions for this user and to keep track of its password, since we need this
password to be aligned with the monitoring config.
The customer_users variable, on the other hand, should be used for users that you
don't care about what happens to them once they're created. This is useful, for example,
if a customer asks you to create a cluster with a initial set of users and you
shouldn't touch these users anymore after they are created.