ansible-scylla-node: Add cluster_local_files_path var
Currently, the node role will look for the file scyllamgr_auth_token.txt in the
same path where the playbook including this role is located.
This might become a problem if the user have playbooks from different locations
including the node role.
In order to fix this problem, this patch adds a cluster_local_files_path var as
a base path to be used for any cluster files that need to be persisted locally.