ansible-scylla-node: Add task list for handling encryption at rest keys
This patch adds a task list for managing local keys for encryption at rest.
This task list will be responsible for copying a local key to all the Scylla
nodes.
If the local key exists but is different from an already existing key on the nodes,
the role will fail.
In order to use this task list, the variables _localhost_key_path and _remote_key_path
must be defined.