rule_config.yml.j2: Stop sending emails for info alerts
This patch creates a new receiver 'team-X-mails-general' and sets
this as the default receiver for the root route. This way, any alerts
that don't match any of the sub-routes will be sent to this receiver.
Since this new receiver don't have an email configured, any alerts sent
to it will be ignored. Today, the only alerts that are going to be ignored
are the info alerts, since this is the only alert level which doesn't match
any of the sub-routes.