ansible-scylla-node: Use Scylla API for getting existing tokens instead of nodetool ring
Currently we're using nodetool ring to get the current tokens and iterating over
its output in order to create an ansible fact with all the tokens.
This might be a problem when using ansible verbose mode since nodetool ring prints
one token per line and every time we append a value to the ansible fact the whole
fact will be printed resulting in a very big number of lines printed, specially
for big clusters.
This patch fixes this problem by using the API to get the current tokens and by writing
them to a file instead of to an ansible fact.