replace_node.yml: Set silences in alertmanager before replacing the node
This patch sets silences in 2 different situations:
  * In the node being replaced, when performing a live node replacement,
    since we know the node we're stopping will start to trigger alerts as
    soon as we stop it and these alerts will only stop once we update
    monitoring with the new node.
  * In the node being replaced and in the new node right after the
    'Prepare nodes before replacement' play is done, since only after this play succeeds
    we know that we're actually good to start due to the validations performed in it.

This patch also expires the silences created once the replacement operations suceeds.

Note that if you're performing a dead node replacement and because of that this node
was already triggering alerts, these alerts will continue until the 'Prepare nodes
before replacement' play finishes, so you might still have to manually set a silence
for the node being replaced.