replace_node.yml: Use one of the alive nodes as a temporary seed during replace
So far we've been expecting the user to provide a list of seeds to be used during
the replace. In this patch, this requirement is being removed and we're using
the first alive node as a temporary seed that will be used only during the replace.
Once the replace is finished, the original list of seeds is restored.