replace_node.yml: Validate seeds and update nodes
In e42ec43000e73054649b11bc5883e2821a74da63 we wrongly removed the task
validating the seeds and also stopped calling the node role for the alive nodes.
However, we need to do both in order to make sure that the replaced node is
not part of the seeds anymore and also to make sure that after the replace
all the nodes have the correct seeds.