ansible-scylla-node: Update gpg keys before running apt update
The task installing gnupg2 runs the equivalent to `apt update` before moving forward
due to `update_cache` being set to `true`. However if this role is being executed for
an old cluster which has an expired key, the `apt update` will fail.
In order to prevent that, this patch does the following:
  * Removes keyring file only if gnupg2 is already installed
  * Removes scylla repo keys only if gnupg2 is already installed
  * Installs gnupg2 only if it's not installed before setting up the keys