scylla-node: fix scylla_setup call
Add missing --no-cpuscaling-setup and --no-fstrim-setup (we explicitly setup these later in the play)
and make sure to always call `scylla_setup` - with the parameters we call it
it should be safe to call it even when scylla server is already running.

We also don't need to give '--nic {{ scylla_nic }} --setup-nic-and-disks' parameters
because they are canceled by --no-sysconfig-setup at the end of the command line
and because we are going to call scylla_sysconfig_setup later in the play explicitly.

We also needed to parametrize the memory locking setup to allow molecule run in containers
on the same VM.

Fixes #98
Fixes #339

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>