scylla-node: rework systemd and perftune related tweaking
Scylla ships an up to date peftune.py for quite some time.
Let's remove the bits that were working around a very old
breakage in it.

With this patch a Role will rely on an auto-selection logic by default.
And one can enforce a specific CPU set using `cpuset_command` parameter.

For example: scylla_cpuset_setup --cpuset 1-7,9-15

In addition to that we need to setup cpuset.conf BEFORE
we run iotune because scylla_io_setup is going to
use the CPUSET value from it.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>