ansible-scylla-node: Get addresses for nodes that are not in play_hosts
In order to run the token_distributor script in the generate_tokens task, we
need to have broadcast_address and rpc_address defined for all the already
bootstrapped nodes, including the ones that are not in the 'play_hosts' variable.
So far we've been assuming that the variables scylla_broadcast_address and
scylla_rpc_address would always be defined for all the nodes, but the user
has no obligation of passing variables for nodes that are not in 'play_hosts',
i.e.: hosts that were excluded from the playbook execution by the usage of
--limit option.
Having that in mind, this patch is defining these variables for bootstrapped nodes
that are not in 'play_hosts' by using the values defined in their 'scylla.yaml' files.
This patch assumes that any already bootstrapped node will have a scylla.yaml
in /etc/scylla with rpc_address and broadcast_address defined.