wipe_node.yml: Remove configuration files
Removing configuration files is necessary for this playbook, since we want
to completely wipe a scylla node in order to be able to re-use the same
instance later.
Since the package module doesn't provide an option for doing that, this patch
is removing its usage and using apt instead. Support for redhat should be added
in the future.