scylla-node/Debian_install.yml: Don't check if the installed scylla version appears in the repo exactly once unless we are upgrading
We have a check that validates that a value in `scylla_version` matches exactly one (full) scylla verison
from the repository so that the Role can install it.

However this check is only relevant when we are going to install anything, and when scylla is already
installed and we are not upgrading - we are not going to. Hence, this check is only making noise in such a case.

One common example of such a situation is when a custom build is installed.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>