replace_node.yml: Move the task that gets the host ids
So far we've been using the host ids only to set the replace_node_first_boot,
but in the next patches we'll also be using them to set nodes that should be ignored.
Having that in mind, let's move this task to outside the block responsible for setting
replace_node_first_boot.