replace_node.yml: Add a 'dead_nodes_to_ignore' variable
Scylla has an 'ignore_dead_nodes_for_replace' config which allows the user
to ignore dead nodes during the replace and so far it was not supported
by this playbook.
This patch adds support to this feature by adding a variable 'dead_nodes_to_ignore'.
This variable should contain a list with the inventory hostnames of the nodes
that should be ignored during the replace.