ansible-scylla-node: Add a way to set additional scylla command line parameters (e.g. seastar ones)
Before this patch we could have set additional scylla command line parameters in a hackish way via io_conf
variable.
This patch adds an explicit way to set SCYLLA_ARGS macro in /etc/default/scylla-server file.

A new variable is added to achieve so: scylla_args_value.
The non-empty value set to this variable is going to override/enforce the SCYLLA_ARGS.

Tested manually with value set and not set.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>