ansible-scylla-node: Don't allow user to provide certs for only a subset of nodes
If the user provides certificates for at least one node, then the user should either
provide certificates for all the other nodes or provide the CA used, otherwise the role
would generate its own CA which would of course be different than the one used by the user.

The main change in this patch is that now the user is allowed to provide certs for the nodes
without having to provide a CA, as long as the certificates are provided for all the nodes.