ansible-scylla-node: Allow using a custom path to the certificates
So far this role allowed the user to provide a parent folder (localhost_cert_path)
for the certificates but would always create the certificates and keys in `{{ localhost_cert_path }}/ssl/{{ hostname }}`,
with fixed names {{ hostname }}.crt and {{ hostname }}.pem,
what prevented us from being able to use a single certificate for all the nodes without having to create
a folder for every single node in the cluster with a same crt and key files.

With the changes in this patch, the 'localhost_cert_path' will indicate the complete path to the crt file and
a variable 'localhost_cert_key_path' was also added, this way the user can save the key in a different path, if necessary.

By having a 'localhost_cert_path' which is not a function of '{{ hostname }}', the user will now be able to have a single
certificate for the whole cluster.

The patch also tries to use the old defaults in the ssl task in case 'localhost_cert_path' and 'localhost_cert_key_path'
are not set in ordre to still support clusters that were relying on the old defaults.