ansible-scylla-node: Add a variable for the truststore in scylla_ssl
So far we were considering that the CA would always be used as a truststore, but
since the CA is not mandatory, this wouldn't always work.
From now on, let's have a dedicated file for the truststore and use the CA as
truststore only if the CA exists and if a dedicated truststore is not given.