async_extra/plugins: always cleanup the state of the job when requested
Before the 'cleanup: True' semantics was rather weird: it wasn't cleaning
the state if the job failed and requiring the operator to cleanup the state
manually. This was not the most convenient user experience.

Let's instead always clean the job alias leaving the actual job status file
on disk if the job execution failed. This kills two birds in a single shot:
  * We can both debug the failed job if we want to.
  * We don't need to do any special handling if simply want to re-run the job.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>