async_extra/tests: Remove unnecessary blocks in a few tests and add proper cleanup before each test
Tests were doing "loops in the air" trying to overcome the original
broken semantics of 'async_task' that was unable to handle failures of
previous executions of the task with the same alias.

Since 'async_task' cleans up the old job alias by itself, now all this complexity is
no longer required.

Make sure to kill the job with a 'blah' alias before running each test
that is going to create a job with this alias.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>