scylla-manager: allow re-use of certain scylla-node defaults
The inclusion of scylla-node role was not optimal, when certain variables were not specified in scylla_manager_db_vars, for example the scylla_repo_key_server, then it resulted in an empty value. Instead, the more appropriate defaults of the scylla-node role should be used.

In addition, there is no need to be wrapping any variables passed to scylla-node role in a scylla_manager_db_vars. Any variable present is already passed as-is to the role. Therefore, we remove this nesting.