replace_node.yml: get the SM cluster name from SM variables
Scylla Manager Role receives clusters information in a 'scylla_clusters'
list. In particular cluster names are stored in scylla_clusters[]['cluster_name'].

Instead, the playbook was using a Scylla Node Role's variable 'scylla_cluster_name' in order
to issue operations with Scylla Manager while the same SM can manage multiple Scylla Clusters
to begin with. And secondly, the `cluster_name` as in 'scylla.yaml' doesn't have to be the same
as the "Cluster name" used by Scylla Manager for the same cluster.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>