ansible-scylla-common role: disable firewall and flush iptables
Each role requires the ability to disable firewall services and flush the iptables.
Due to this commonality between roles, a new role `ansible-scylla-common` is introduced.
This common role currently only features the ability to disable the firewall.

Fixes #470

The parameter `firewall_enabled` has also been deprecated, as this gives the impression that
the firewall will be enabled, which is not true. Instead, a more semantically correct variable
has been introduced `disable_firewall` which when set to true will disable the firewall, and if not,
no action is taken.

Signed-off-by: Vincent Reniers <vincent.reniers@scylladb.com>