ansible-scylla-node: upgrade/main.yml: cohesive and improved validation checks that are skippable
An upgrade from 2024.1 to 2025.1 is not possible, because it will fail on validation. In addition, we want to allow more flexibility on skipping validation.

In addition, the task attempts to auto-correct when upgrading to a too high of a major version. This autocorrect behavior can be unexpected, and sometimes not needed (e.g. when going to 2024.1 to 2025.1). We rather execute what a user has explicitly given as input, and if not fail, or allow a skip the validation when desired. Therefore, we remove `upgrade_autocorrect_version`.
In addition, we group the upgrade validation logic, and allow for a validation skip via `upgrade_skip_validation` and `upgrade_skip_upgrade_path_validation`. The support matrix check can now also be skipped.

We remove `upgrade_force` which is not quite semantically clear. In practice, it only skips the check whether a major or minor upgrade is effectively doing such an upgrade. This should in general done, however we now allow this to be skipped if needed via `upgrade_skip_validation`.

We remove the internal `upgrade` flag which as part of `scylla_upgrade`, and indicated whether it was an upgrade or downgrade. It was only used to fail in case it was a not permitted downgrade, or to autocorrect the version to a lower or higher version respectively. Since we removed autocorrect, and since this downgrade is not a well implemented feature, we remove it this internal flag.