ansible-scylla-common: set apt config file to prevent dpkg lock timeout
In a previous PR (#457), we had set module defaults to increase the apt timeout lock.
However, it seems this still doesn't prevent issues when an `unatttend-upgr` executes, which will claim a dpkg lock.
This is due to a known Ansible bug in the `apt` module, which is supposedly resolved in version `2.19`.
In order to workaround this issue for now, we allow an apt configuration file to be set which configures a timeout.

Signed-off-by: Vincent Reniers <vincent.reniers@scylladb.com>