ssl cert generation: use SAN instead of deprecated common name
We use subject alternative names (SAN) instead of the deprecated common name.
In case of monitoring, we also alter the SAN that is set. A common name can be IP or DNS, whereas it has to be rather specific in case of a SAN.
A rather good default is chosen which is the `ansible_fqdn`, although in the future we can possibly also make it a configurable.