ansible-scylla-node: allow to specify SAN entries in certificates
The ScyllaDB node when using SSL responds with it's certificate, and in case of Alternator the endpoint that is used should match with what is present in the certificate. Previously, we used a default that is not used as endpoint. This patch addresses this, and sets good defaults as the subject alternative names (SAN) in the certificate. In addition, we allow a configurable that specifies additional DNS entries.