replace_node.yml: Skip scylla.yaml changes and scylla startup if scylla is already running
If scylla is already running in the new node we assume that it was started by a previous execution
of this playbook and thus there is no point in changing scylla.yaml or starting scylla again.
In order to skip all the tasks that change scylla.yaml and start scylla, this patch groups them together
in a single block.