ansible-scylla-node: Copy remote encryption-at-rest keys to localhost
Currently, this role only copies keys from localhost to the remote nodes.
This requires users to either manually generate the keys or manually copy them from
the remote nodes in order to manage keys with this role.
Since Scylla automatically generates the keys on the remote nodes after bootstrap,
this patch updates the role to also copy those keys back to localhost.