molecule.yml: Set cgroupns_mode: host
The self-hosted runner was recently re-deployed with a more recent docker version that
has cgroupsv2 enabled by default.
As a result, our molecule tests started to fail with errors like:
"Failed to connect to bus: No such file or directory" when a task attempts to run a systemd command.

This seems to be caused by docker not fully supporting systemd inside containers, as discussed in these issues:
https://github.com/moby/moby/issues/42275
https://github.com/systemd/systemd/issues/19245
https://github.com/ansible/molecule/issues/3632

In order to workaround this problem, this patch sets 'cgroupns_mode: host', adjusting the container
to use the host's cgroup namespace.